Module_googlevision/requirements.txt
flask
flask_cors
flask-restful
google-api-python-client
google-auth-httplib2
google-auth
requests